refactoring, fix lint issues
