+++ src/consumer/versions-ops/merge-version/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+export { default as threeWayMerge } from './three-way-merge';
+export { mergeVersion, FileStatus, getMergeStrategy, MergeOptions } from './merge-version';
+export type { MergeStrategy, ApplyVersionResults, ApplyVersionResult } from './merge-version';
